Disable active_record.belongs_to_required_by_default
When Rails 5.1 was released belongs_to requires by default that the
related record exists. When the relation is we have to add
optional: true

Unfortunately this has the side effect of adding errors to the
relationship instead of the foreign key field[0]. So, if we have:

```
class Subject < ActiveRecord::Base
  belongs_to :teacher
end
```

and in our form, something like:

```
<%= govuk_collection_radio_buttons :teacher_id, @teachers, :id, :name %>
```

If no teacher is selected the error will be added to subject on the
`teacher` field, which isn't present in the form, leading to a mismatch
between the `govuk_error_summary` and the form field. The radio
collection won't be highlighted in red and the link from the error
summary won't match the id of the first radio button input.

I suggest we disable this setting from the beginning and add validation
explicitly.

[0] https://github.com/rails/rails/pull/46494