Add page_data helper to standardise page setup
When we have lots of pages, ensuring they are consistently built becomes
more of a challenge. Moving as much to the layout as possible helps in
this regard, so this change introduces the page_data helper which sets
the following things (so far):

* page_title  - the page title follows the regular format of the title of
                the current page followed by the service name and GOV.UK.
                When `errors: true` the title will be prefixed with
                `Error: `[0]. It is expected that this is usually set to
                `errors: @the_form_object.errors.any?`
* page_header - this defaults to the title text but can be overridden if
                the two differ. The size is controllable using
                `header_size`

We can extend this in the future if needed, allowing the layout to be
overridden[1] might be useful.

[0] https://design-system.service.gov.uk/components/error-summary/#how-it-works
[1] https://guides.rubyonrails.org/layouts_and_rendering.html#choosing-layouts-at-runtime