Fix a unpredictable test
The induction period factory uses a sequence to generate the start and
finish dates and when the sequence is high and the finished_on is
hardcoded there's a chance the finished_on is before the started_on.

Now we're generating ranges this causes an error at the database level
(good) and causes a failure.