Add validation context for terms/induction type
The pending_induction_submission model will be used for both bulk and
single ECTs so we need to adjust it so that it can be populated all at
once or sequentially.

It will always need some details, like appropriate body and TRN but
when registering a single ECT we'll identify/record the teacher first so
will only be storing teachr-related stuff at that point. Later when we
add the induction periods we can use the :record_period context to
ensure that part of the record is valid.