Add find ECT page and stubbed service
Finding an ECT is the first stage of AB ECT registration. The TRN and
date of birth will be searched for in the TRS API.

When there is a match, we store the relevant bits of the result in the
PendingInductionSubmission record and then move to the next stage where
the AB user confirms they have the right teacher.