Add features covering the AB claim flow
This commit also introduces the TRS::FakeAPIClient. It's only used in a
basic sense now but we will be able to extend it so it responds in
different ways should the need arise.

It might also make sense to use this in development mode if an API key
isn't set, so it's easier to get to pages on the other side of checks.