Add a teachers show action for ABs
This lets an AB view a teacher's record. It's placeholder for now but we
can add buttons to support the upcoming transactions.

We also introduce Teachers::Name which will help us build the right name
contextually. In most places we will just want either the 'corrected
name' if it's set, and if not fall back to the first name + last name
from TRS.

There may be some places in the future where we want to show both names,
like "Bob Smith (Robert Smith)" or similar too.