Make searching work with accented strings
The way we're searching uses a `search` field in the teachers table that
contains the concatenated first_name, last_name and corrected_name
fields in a tsvector (for full text indexing).

Unfortunately, this didn't work with accented characters like 'André';
we'd expect to be able to find that record with 'Andre'.

Initially we wanted to simply use PostgreSQL's unaccent[0] module to
standardise the text. This ultiately didn't work because unaccent is
stable but not immutable[1]

Instead of taking the easy but slow route of just doing the conversions
on the fly, this approach is a little more involved.

It won't make any difference in deployed environments where we always
migrate rather than repeatedly drop/create/seed the database like we do
in dev.

We can solve the problem by adding a new text search confuguration that
applies the unaccent module, and use it to build the tsvector we index
on.

Without switching from schema.rb to schema.sql, we can't do this as part
of the migration but just need to make sure it's set up when the
database is built.

The cleanest way to do this is to add a rake task that enhances[2] the
db:schema_load task by running the snippet of SQL before it. This
ensures what we need is in place before creating the tables which depend
on it.

[0] https://www.postgresql.org/docs/current/unaccent.html
[1] https://www.postgresql.org/message-id/flat/201012021544.oB2FiTn1041521@wwwmaster.postgresql.org#201012021544.oB2FiTn1041521@wwwmaster.postgresql.org
[2] https://www.rubydoc.info/gems/rake/Rake/Task#enhance-instance_method