Add enable_schools_interface config option
This configuration option sets whether visitors to the app can see the
school functionality. We need to add it because the appropriate body
functionality will be live before the schools, but eventually the
schools bit will be the main bit.

So, for now, we will redirect people in sandbox/production to
/appropriate-body, but in all other envs the schools landing page will
be visible on /

The setting can be changed by setting the ENABLE_SCHOOLS_INTERFACE
environment variable.

Refs #429