Use persona based login for AB tests
The early tests were written when we thought appropriate body users
would sign in using an email address and magic link. We now know they'll
be using DfE Sign-in so we're best switching the tests over to using
personas.

We'll identify appropriate body users by checking for the presence of an
appropriate_body_id in the session.

I've added a screen with buttons to make choosing/switching the current
AB easier. It's only intended to be used by the tests so there's no
links to it etc - perhaps if it's useful we'll add these later.