Add the Sessions::SessionUser class
We currently have three different kinds of user who can log in; DfE
Sign-in, OTP (admin) and Persona (test) and no consistent way of
interacting with them.

This change adds the Sessions::SessionUser class which aims to do that
by providing a standard interface - each of the 3 kids of user records
can be used to initiate it and it provides a set of methods to query it
appropriately.

Co-authored-by: tonyheadford <tony@binarycircus.com>